Consolidate database into a single dataset

This function collapses a set or database of (q)datasets into a single dataset with some combination of the rows, columns, and observations of the parent datasets. The function includes separate arguments for the rows and columns, as well as for how to resolve conflicts in observations across datasets. This provides users with considerable flexibility in how they combine qData. For example, users may wish to stick to units that appear in every dataset but include variables coded in any dataset, or units that appear in any dataset but only those variables that appear in every dataset. Even then there may be conflicts, as the actual unit-variable observations may differ from dataset to dataset. Here we (will) offer a number of resolve methods that enable users to choose how conflicts between observations are resolved.

consolidate(
  database,
  rows = c("any", "every"),
  cols = c("any", "every"),
  resolve = "coalesce",
  key = "qID"
)

Arguments

database
A qPackage database object

rows
Which rows or units to retain. By default "any" (or all) units are retained, but another option is "every", which retains only those units that appear in all parent datasets.

cols
Which columns or variables to retain. By default "any" (or all) variables are retained, but another option is "every", which retains only those variables that appear in all parent datasets.

resolve
How conflicts between observations should be resolved. Currently only "coalesce" is offered, which takes the first non-NA value, but options should soon include "max", "min", "mean", "mode" and "median", as well as "append".

key
An ID column to collapse by. By default "qID".

Value

A single tibble/data frame.

See also

Examples

pluck(emperors, "UNRV")

#> # A tibble: 99 x 7
#>    ID               Beg       End       Birth Death FullName          Dynasty   
#>    <chr>            <messydt> <messydt> <chr> <chr> <chr>             <chr>     
#>  1 Augustus         0027 BC   0014 BC   63 BC 14    Gaius Julius Cae~ Julio-Cla~
#>  2 Tiberius         0014 BC   0037      42 BC 37    Tiberius Claudiu~ Julio-Cla~
#>  3 Gaius (Caligula) 0037      0041      12    14    Gaius Caesar Ger~ Julio-Cla~
#>  4 Claudius         0041      0054      10 BC 41    Tiberius Claudiu~ Julio-Cla~
#>  5 Nero             0054      0068      37    68    Claudius Nero Ca~ Julio-Cla~
#>  6 Galba            0068      0069      3 BC  69    Servius Sulpiciu~ Year of t~
#>  7 Otho             0069      0069      32    69    Marcus Salvius O~ Year of t~
#>  8 Vitellius        0069      0069      15    69    Aulus Vitellius ~ Year of t~
#>  9 Vespasian        0069      0079      9     79    Titus Flavius Ve~ Year of t~
#> 10 Titus            0079      0081      39    79    Titus Flavius Ve~ Flavian   
#> # ... with 89 more rows

consolidate(emperors, "any", "any", key = "ID")

#> # A tibble: 137 x 15
#>    ID        CityBirth  ProvinceBirth Rise  Cause Killer Era   Notes Verif Birth
#>    <chr>     <chr>      <chr>         <chr> <chr> <chr>  <chr> <chr> <chr> <chr>
#>  1 Augustus  Rome       Italia        Birt~ Assa~ Wife   Prin~ birt~ Redd~ 0062~
#>  2 Tiberius  Rome       Italia        Birt~ Assa~ Other~ Prin~ birt~ Redd~ 0041~
#>  3 Caligula  Antitum    Italia        Birt~ Assa~ Senate Prin~ assa~ Redd~ 0012~
#>  4 Claudius  Lugdunum   Gallia Lugdu~ Birt~ Assa~ Wife   Prin~ birt~ Redd~ 0009~
#>  5 Nero      Antitum    Italia        Birt~ Suic~ Senate Prin~ NA    Redd~ 0037~
#>  6 Galba     Terracina  Italia        Seiz~ Assa~ Other~ Prin~ birt~ Redd~ 0002~
#>  7 Otho      Terentinum Italia        Appo~ Suic~ Other~ Prin~ NA    NA    0032~
#>  8 Vitellius Rome       Italia        Seiz~ Assa~ Other~ Prin~ NA    NA    0015~
#>  9 Vespasian Falacrine  Italia        Seiz~ Natu~ Disea~ Prin~ NA    NA    0009~
#> 10 Titus     Rome       Italia        Birt~ Natu~ Disea~ Prin~ NA    NA    0039~
#> # ... with 127 more rows, and 5 more variables: Death <chr>, FullName <chr>,
#> #   Dynasty <chr>, Beg <messydt>, End <messydt>

consolidate(emperors, "every", "every", key = "ID")

#> # A tibble: 41 x 3
#>    ID        Beg           End       
#>    <chr>     <messydt>     <messydt> 
#>  1 Augustus  0026-01-16 BC 0014-08-19
#>  2 Tiberius  0014-09-18    0037-03-16
#>  3 Claudius  0041-01-25    0054-10-13
#>  4 Nero      0054-10-13    0068-06-09
#>  5 Galba     0068-06-08    0069-01-15
#>  6 Otho      0069-01-15    0069-04-16
#>  7 Vespasian 0069-12-21    0079-06-24
#>  8 Titus     0079-06-24    0081-09-13
#>  9 Domitian  0081-09-14    0096-09-18
#> 10 Nerva     0096-09-18    0098-01-27
#> # ... with 31 more rows