Find and download packages in the qData ecosystem

Find and download packages in the qData ecosystem

get_packages(pkg)

Arguments

pkg
A character vector of package names or number of a package

Value

If no package name is provided, this function prints a table (tibble) to the console with details on packages that are currently available within the qData ecosystem. This includes the name and description of the package, the latest installed and release version number, and the latest release date, and a string of contributors. It also include a list of numbers which orders the package and can be used to load the respective package instead of the name. If one or more package names are provided, these will be installed from Github.

Details

The function finds and download other packages that belong to the qData ecosystem of data packages. It allows for users to rapidly access the names and other descriptive information of these packages by simply calling the function. If users intend to download a package from the ecosystem, they can to type the package name within the function.

Examples

get_packages()

#> Warning:  1 failed to parse.

#> # A tibble: 5 x 6
#>   name     full_name         
#>   <chr>    <chr>             
#> 1 qCreate  globalgov/qCreate 
#> 2 qData    globalgov/qData   
#> 3 qEnviron globalgov/qEnviron
#> 4 qStates  globalgov/qStates 
#> 5 qTrade   globalgov/qTrade  
#>   description                                             installed latest    
#>   <chr>                                                   <chr>     <chr>     
#> 1 Support for creating new qPackages                      0.1.1     0.1.0     
#> 2 An R portal for ensembled global governance data        0.5.0     0.5.0     
#> 3 qPackage for ensembled data on environmental agreements 0.0.3     0.0.3     
#> 4 qPackage for ensembled data on sovereign states         0.0.5     0.0.4     
#> 5 qPackage for ensembled data on trade agreements         0.0.1     Unreleased
#>   updated   
#>   <date>    
#> 1 2021-07-16
#> 2 2021-07-26
#> 3 2021-08-03
#> 4 2021-08-19
#> 5 NA