Standardises a wide range of of date inputs

The function standardises a wide range of date inputs parsed through it. It accepts date inputs in different formats, incomplete dates, historical dates and future dates. It also creates nested vectors of dates for vague date inputs, ambiguous and ranged dates, into a range of dates.

standardise_dates(...)

Arguments

...
One (ymd) or three (yyyy, mm, dd) variables

Value

Nested vector of POSIXct dates that includes a range of dates

Details

The function seeks to convert a wide range of dates into dates so that these can be meaningfully used for analysis. There are several limitations of other date wrangling packages and/or functions for dealing with incomplete dates, dates with different or inconsistent formats or historical dates for which standardise_dates() can be used. It also converts ambiguous and ranged dates into a range of dates. The function allows only for dmy or ymd date formats at present, since mdy may introduce errors.

Examples

dates_comparison <- tibble::tribble(~Example, ~OriginalDate,
"A normal date", "2010-01-01",
"A historical date", "1712-01-01",
"A really historical date", "712-01-01",
"A very historical date", "012-01-01",
"A clearly future date", "9999-12-31",
"A not so clearly future date", "2599-12-31")
dates_comparison %>% dplyr::mutate(
lubridate = suppressWarnings(lubridate::as_date(OriginalDate)),
anytime = anytime::anydate(OriginalDate),
qData = qData::standardise_dates(OriginalDate)
) %>% print(n = 25)

#> # A tibble: 6 x 5
#>   Example                      OriginalDate lubridate  anytime    qData     
#>   <chr>                        <chr>        <date>     <date>     <date>    
#> 1 A normal date                2010-01-01   2010-01-01 2010-01-01 2010-01-01
#> 2 A historical date            1712-01-01   1712-01-01 1712-01-01 1712-01-01
#> 3 A really historical date     712-01-01    NA         NA         0712-01-01
#> 4 A very historical date       012-01-01    NA         NA         0012-01-01
#> 5 A clearly future date        9999-12-31   9999-12-31 9999-12-31 9999-12-31
#> 6 A not so clearly future date 2599-12-31   2599-12-31 2599-12-31 9999-12-31