Find and download packages in the qData ecosystem

Find and download packages in the qData ecosystem

get_packages(pkg)

Arguments

pkg
A character vector of package names or number of a package

Value

If no package name is provided, this function prints a table (tibble) to the console with details on packages that are currently available within the qData ecosystem. This includes the name and description of the package, the latest installed and release version number, and the latest release date, and a string of contributors. It also include a list of numbers which orders the package and can be used to load the respective package instead of the name. If one or more package names are provided, these will be installed from Github.

Details

The function finds and download other packages that belong to the qData ecosystem of data packages. It allows for users to rapidly access the names and other descriptive information of these packages by simply calling the function. If users intend to download a package from the ecosystem, they can to type the package name within the function.

Examples

get_packages()

#> # A tibble: 0 × 1
#> # … with 1 variable: updated <date>

get_packages("qStates")

#> Using github PAT from envvar GITHUB_PAT

#> Downloading GitHub repo globalgov/qStates@HEAD

#> slam       (NA -> 0.1-48) [CRAN]
#> data.table (NA -> 1.14.0) [CRAN]
#> dtt        (NA -> 0.1-2 ) [CRAN]
#> zoo        (NA -> 1.8-9 ) [CRAN]
#> NLP        (NA -> 0.2-1 ) [CRAN]
#> textshape  (NA -> 1.7.3 ) [CRAN]
#> syuzhet    (NA -> 1.0.6 ) [CRAN]
#> qdapRegex  (NA -> 0.7.2 ) [CRAN]
#> mgsub      (NA -> 1.7.2 ) [CRAN]
#> lexicon    (NA -> 1.2.1 ) [CRAN]
#> english    (NA -> 1.2-5 ) [CRAN]
#> textclean  (NA -> 0.9.3 ) [CRAN]

#> Installing 12 packages: slam, data.table, dtt, zoo, NLP, textshape, syuzhet, qdapRegex, mgsub, lexicon, english, textclean

#> Installing packages into ‘/Users/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)

#> 
#> The downloaded binary packages are in
#> 	/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//Rtmpys8kVA/downloaded_packages

#> Running `R CMD build`...

#> * checking for file ‘/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpys8kVA/remotes1253651b54ca/globalgov-qStates-352455f/DESCRIPTION’ ... OK
#> * preparing ‘qStates’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building ‘qStates_0.0.3.tar.gz’

#> Installing package into ‘/Users/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)

#> Loading required package: pointblank